Refactor: Extract OnThisPage component and use in blog and docs
